requirement.txt
manim
streamlit
numpy
matplotlib.pyplot
